refactor: Move types to `/types/`
