+++ src/app/tournament/[uuid]/_components/DeletePlayerModal.tsx
@@ -5,7 +5,7 @@ import { Modal } from "@/elements/Modal";
 import { Notification } from "@/elements/Notification";
 import { RoundedListbox } from "@/elements/RoundedListbox";
 import { usePlayers } from "@/hooks/usePlayers";
-import type { Player } from "@/lib/types";
+import type { Player } from "@/types/app.types";
 
 type DeletePlayerModalProps = {
   isOpen: boolean;
