+++ src/app/tournament/[uuid]/_components/EditPlayerModal.tsx
@@ -6,8 +6,8 @@ import { Modal } from "@/elements/Modal";
 import { Notification } from "@/elements/Notification";
 import { RoundedListbox } from "@/elements/RoundedListbox";
 import { usePlayers } from "@/hooks/usePlayers";
-import type { Player } from "@/lib/types";
 import { parseFormString } from "@/lib/utils";
+import type { Player } from "@/types/app.types";
 
 type EditPlayerModalProps = {
   isOpen: boolean;
