+++ src/app/tournament/[uuid]/_components/TableSeat.tsx
@@ -7,8 +7,13 @@ import { usePlayers } from "@/hooks/usePlayers";
 import { useTables } from "@/hooks/useTables";
 import { useTournaments } from "@/hooks/useTournaments";
 import { getPointDeltas } from "@/lib/scoring";
-import type { Player, PointsAnimationEvent, Table, Wind } from "@/lib/types";
 import { useTournamentContext } from "@/providers/TournamentProvider";
+import type {
+  Player,
+  PointsAnimationEvent,
+  Table,
+  Wind,
+} from "@/types/app.types";
 import { WinSelector } from "./WinSelector";
 
 type TableSeatProps = {
