+++ src/app/tournament/[uuid]/sessions/_components/ViewGraphModal.tsx
@@ -1,8 +1,8 @@
 import { useMemo } from "react";
 import { BarChart } from "@/elements/BarChart";
 import { Modal } from "@/elements/Modal";
-import type { Player } from "@/lib/types";
 import { rankPlayers } from "@/lib/utils";
+import type { Player } from "@/types/app.types";
 
 type ViewGraphModalProps = {
   players: Player[];
